+++ src/XUnity.AutoTranslator.Plugin.Core/Constants/PluginData.cs
@@ -11,6 +11,6 @@ namespace XUnity.AutoTranslator.Plugin.Core.Constants
 
       public const string Name = "XUnity Auto Translator";
 
-      public const string Version = "2.15.3";
+      public const string Version = "2.15.4";
    }
 }